Remove legacy hawkRequest
