+++ src/lib/hawk-request.js
@@ -1,98 +1,92 @@
-/**
- * This is the original rp version of hawk request. This should be deprecated
- * but seems more stable.
- */
-const request = require('request')
 const Hawk = require('hawk')
 const config = require('../config')
-
-const HTTP_OK = 200
+const request = require('./request')
 
 function getHawkHeader(credentials, requestOptions) {
   if (config.isTest) {
     return 'hawk-test-header'
   }
 
-  const { uri, method } = requestOptions
+  const { url, method } = requestOptions
 
   // Generate Authorization request header
   // Ensure backend is using same protocol for hash generation
-  return Hawk.client.header(uri, method, {
+  return Hawk.client.header(url, method, {
     credentials,
     payload: '',
-    contentType: '',
+    contentType: 'application/json',
   })
 }
 
-function createPromiseRequest(
-  requestOptions,
-  credentials,
-  clientHeaderArtifacts
-) {
-  return new Promise((resolve, reject) => {
-    request(requestOptions, (err, response, responseBody) => {
-      if (err || response?.statusCode !== HTTP_OK) {
-        return reject(new Error(err))
-      }
-      if (!config.isTest) {
-        let isValid = false
-        try {
-          // Authenticate the server's response must use raw response body here
-          isValid = Hawk.client.authenticate(
-            response,
-            credentials,
-            clientHeaderArtifacts,
-            { payload: responseBody }
-          )
-        } catch (e) {
-          const err = new Error('Unable to validate response')
-          err.rootError = e
-          return reject(err)
-        }
-
-        if (!isValid) {
-          return reject(new Error('Invalid response'))
-        }
-      }
-      const body = JSON.parse(responseBody)
-
-      const statusCode = response.statusCode
-      if (statusCode >= 200 && statusCode < 300) {
-        return resolve(body)
-      }
-      const error = new Error(
-        `Got a ${statusCode} response code for ${requestOptions.uri}`
-      )
-      error.responseBody = body
-
-      return reject(error)
-    })
+async function hawkRequest(requestOptions, credentials, clientHeaderArtifacts) {
+  const response = await request({
+    validateStatus: (status) => status >= 200 && status < 500,
+    ...requestOptions,
   })
+
+  if (!config.isTest) {
+    let isValid = false
+    const payload = JSON.stringify(response.data || {})
+    try {
+      // Authenticate the server's response must use raw response body here
+      isValid = Hawk.client.authenticate(
+        response,
+        credentials,
+        clientHeaderArtifacts,
+        payload
+      )
+    } catch (e) {
+      const err = new Error('Unable to validate response')
+      err.rootError = e
+      throw err
+    }
+
+    if (!isValid) {
+      throw new Error('Invalid response')
+    }
+  }
+
+  if (response.status >= 300) {
+    const error = new Error(
+      `Got a ${response.status} response code for ${requestOptions.uri}`
+    )
+    error.responseBody = response.data
+    throw error
+  }
+
+  return response.data
+}
+
+function hawkRequestPromise(...params) {
+  return new Promise((resolve, reject) =>
+    hawkRequest(...params)
+      .then((response) => resolve(response))
+      .catch((error) => reject(error))
+  )
 }
 
 async function sendHawkRequest(
-  uri,
-  credentials = config.hawkCredentials.dataHubBackend
+  url,
+  credentials = config.hawkCredentials.dataHubBackend,
+  timeout
 ) {
-  if (!uri) {
-    throw new Error('Uri is required')
+  if (!url) {
+    throw new Error('Url is required')
   }
   const requestOptions = {
-    uri,
+    url,
     method: 'GET',
     headers: {
       accept: 'application/json',
+      'Content-Type': 'application/json',
     },
   }
 
   const clientHeader = getHawkHeader(credentials, requestOptions)
-  requestOptions.headers.Authorization = clientHeader.header
+  requestOptions.headers.Authorization = clientHeader.header || clientHeader
+  timeout && (requestOptions.timeout = timeout)
 
-  return createPromiseRequest(
-    requestOptions,
-    credentials,
-    clientHeader.artifacts
-  )
+  return hawkRequestPromise(requestOptions, credentials, clientHeader.artifacts)
 }
 
 module.exports = sendHawkRequest
